Framework of server rounds, minor bug fixes too
Working all day to get the general outline of the server rounds up and running. There's two parts that I'm still working on in terms of the server rounds: forcing clients to only send messages within a round (any messages that are sent outside of the round duration should be pushed to the next round) and computing the new dead drop server once the round finishes. I'll continue working on this in the morning.

I'm just committing what I have now because I need sleep.